Solved bug in hosting processor caused by repository existence checking.
The main cause of the problem was that a method
checkIfRepositoryExistsByName was called to make the checking process,
but the id were passed as parameter instead of the name. So I have
created a new method called checkIfRepositoryExistsById in
RepositoryDocumentHandler, and now this method is used to make the
repository existence checking for HostingService.